🚀 修复(deploy/bash.sh): 优化docker命令，确保服务已启动时才执行停止和重建操作

+++ deploy/bash.sh
@@ -1,6 +1,11 @@
 # bin/bash
 cd /tools
-docker compose down
-docker rmi gsdukbh/tools-web
-docker build -t $IMAGE_TAG -f Dockerfile ./
-docker compose up -d
+if docker compose ps | grep -q 'Up'; then
+    docker compose down
+    docker rmi gsdukbh/tools-web
+    docker build -t $IMAGE_TAG -f Dockerfile ./
+    docker compose up -d
+else
+    docker build -t $IMAGE_TAG -f Dockerfile ./
+    docker compose up -d
+fi
